+++ plotly/src/traces/scatter3d.rs
@@ -154,6 +154,8 @@ where
     /// per-point (the ones that are `arrayOk: true`) are available.
     #[serde(rename = "texttemplate")]
     text_template: Option<Dim<String>>,
+    #[serde(rename = "texttemplatefallback")]
+    text_template_fallback: Option<Dim<String>>,
     /// Sets hover text elements associated with each (x, y, z) triplet. The
     /// same text will be associated with all data points. To be seen, the
     /// trace `hover_info` must contain a "Text" flag.
@@ -182,6 +184,8 @@ where
     /// completely, use an empty tag `<extra></extra>`.
     #[serde(rename = "hovertemplate")]
     hover_template: Option<Dim<String>>,
+    #[serde(rename = "hovertemplatefallback")]
+    hover_template_fallback: Option<Dim<String>>,
     /// Sets the hover text formatting rulefor `x` using d3 formatting
     /// mini-languages which are very similar to those in Python. For numbers, see: <https://github.com/d3/d3-format/tree/v1.4.5#d3-format>. And for
     /// dates see: <https://github.com/d3/d3-time-format/tree/v2.2.3#locale_format>. We add two items to d3's