+++ plotly/src/traces/surface.rs
@@ -145,6 +145,8 @@ where
     hover_label: Option<Label>,
     #[serde(rename = "hovertemplate")]
     hover_template: Option<Dim<String>>,
+    #[serde(rename = "hovertemplatefallback")]
+    hover_template_fallback: Option<Dim<String>>,
     #[serde(rename = "hovertext")]
     hover_text: Option<Dim<String>>,
     #[serde(rename = "legendgroup")]